requirements.txt
numpy
opencv-python
torch
torchvision
pillow
tqdm
scikit-learn
tensorboard